+++ examples/multiFileProject/multiFileProject.cpp
@@ -1,0 +1,14 @@
+/****************************************************************************************************************************
+  multiFileProject.cpp
+  For STM32 with built-in LAN8742A Ethernet (Nucleo-144, DISCOVERY, etc)
+
+  AsyncUDP_STM32 is a Async UDP library for the STM32 using built-in LAN8742A Ethernet
+
+  Based on and modified from ESPAsyncUDP Library (https://github.com/me-no-dev/ESPAsyncUDP)
+  Built by Khoi Hoang https://github.com/khoih-prog/AsyncUDP_STM32
+  Licensed under MIT license
+*****************************************************************************************************************************/
+
+// To demo how to include files in multi-file Projects
+
+#include "multiFileProject.h"